Split DesktopSettingsScreen.kt into section composables
